[fix]間違った型だったので正しい型へ修正 [update]クールダウンを追加

+++ Source/Fairy_Speed.h
@@ -8,7 +8,8 @@ private:
 public:
 	Fairy_Speed() = default;
 	Fairy_Speed(float _speed, float _power, int _durability, eDirection _direction);
-	Fairy_Speed(float _speed, float _power, int _durability, eDirection _direction, int _x, int _y);
+	Fairy_Speed(float _speed, float _power, int _durability, eDirection _direction,
+		float _x, float _y);
 	void Update(CastleManager* _castleManager, BasePlayer* _player, BulletManager* _bulletManager);
 	void Draw();
 };
